requirements.txt
flask
requests
Flask-SocketIO
eventlet
pika
websocket
evdev
mysql-connector-python
python-dotenv
mod-wsgi